+++ app/services/carto/user_metadata_export_service.rb
@@ -125,7 +125,7 @@ module Carto
     def build_user_from_hash(exported_user)
       user = User.new(exported_user.slice(*EXPORTED_USER_ATTRIBUTES - [:id]))
 
-      user.feature_flags_user = exported_user[:feature_flags].map { |ff_name| build_feature_flag_from_name(ff_name) }
+      user.self_feature_flags_user = exported_user[:feature_flags].map { |ff_name| build_feature_flag_from_name(ff_name) }
                                                              .compact
 
       user.assets = exported_user[:assets].map { |asset| build_asset_from_hash(asset.symbolize_keys) }
@@ -330,7 +330,7 @@ module Carto
     def export(user)
       user_hash = EXPORTED_USER_ATTRIBUTES.map { |att| [att, user.attributes[att.to_s]] }.to_h
 
-      user_hash[:feature_flags] = user.feature_flags_user.map(&:feature_flag).map(&:name)
+      user_hash[:feature_flags] = user.feature_flags_names
 
       user_hash[:assets] = user.assets.map { |a| export_asset(a) }